requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Data / metrics
numpy
pandas
scikit-learn
Pillow

# Visualisation
matplotlib

# Demo
gradio==4.44.1
fastapi==0.111.1
starlette==0.37.2
huggingface_hub>=0.20.0

# Experiment tracking (optional — only needed for training)
wandb
